changes for sheepdog and peregrine
