+++ .gitignore
@@ -10,4 +10,3 @@ output/
 gen3-integration-tests/test_data/lock
 gen3-integration-tests/test_data/configuration
 gen3-integration-tests/test_data/graph_data
-gen3-integration-tests/test_data/sheepdog_peregrine
