+++ gen3-integration-tests/pyproject.toml
@@ -13,7 +13,6 @@ include = [
 [tool.poetry.dependencies]
 python = "^3.9"
 allure-pytest = "^2.13.2"
-boto3 = "~1.28.33"
 cdislogging = "<2"
 data-simulator = {git = "https://github.com/uc-cdis/data-simulator", rev = "1.6.1"}
 filelock = "<4"
@@ -27,7 +26,6 @@ pytest-playwright = "^0.4.3"
 pytest-xdist = "^3.3.1"
 python-dotenv = "^1.0.0"
 requests = "^2.31.0"
-playwright = "^1.43.0"
 
 [build-system]
 requires = ["poetry-core"]
