+++ gen3-integration-tests/utils/misc.py
@@ -108,74 +108,3 @@ def retry(times, delay, exceptions):
         return newfn
 
     return decorator
-
-
-def create_program(program_name, auth_header, user="main_account"):
-    auth = Gen3Auth(refresh_token=pytest.api_keys[user])
-    program_form = (
-        '{"name":"'
-        + program_name
-        + '","type":"program","dbgap_accession_number":"'
-        + program_name
-        + '"}'
-    )
-    response = requests.post(
-        url=pytest.root_url + "/api/v0/submission",
-        headers=auth_header,
-        auth=auth,
-        data=program_form,
-    )
-    if response.status_code == 200:
-        logger.info("Successfully Created/Updated program")
-    else:
-        logger.info(response.json())
-        logger.error("Failed to Create/Update program")
-
-
-def create_project(program_name, project_name, auth_header, user="main_account"):
-    auth = Gen3Auth(refresh_token=pytest.api_keys[user])
-    project_form = (
-        '{"type":"project","code":"'
-        + project_name
-        + '","name":"'
-        + project_name
-        + '","dbgap_accession_number":"'
-        + project_name
-        + '","state":"open","releasable":true}'
-    )
-    response = requests.post(
-        url=pytest.root_url + "/api/v0/submission/" + program_name,
-        headers=auth_header,
-        auth=auth,
-        data=project_form,
-    )
-    if response.status_code == 200:
-        logger.info("Successfully Created/Updated project")
-    else:
-        logger.info(response.json())
-        logger.error("Failed to Create/Update project")
-
-        
-def retry(times, delay, exceptions):
-    """
-    Decorator that retries the wrapped function/method `times` times if the exceptions
-    listed in ``exceptions`` are thrown waiting for `delay` seconds between retries
-    """
-
-    def decorator(func):
-        def newfn(*args, **kwargs):
-            attempt = 1
-            while attempt <= times:
-                time.sleep(delay)
-                try:
-                    return func(*args, **kwargs)
-                except exceptions:
-                    print(
-                        f"Errored when trying to run {func}, attempt {attempt} of {times}"
-                    )
-                    attempt += 1
-            return func(*args, **kwargs)
-
-        return newfn
-
-    return decorator
