+++ gen3-integration-tests/utils/test_setup.py
@@ -11,8 +11,10 @@ from datasimulator.main import (
 from pathlib import Path
 
 from services.fence import Fence
-from utils.misc import create_program, create_project
+from services.graph import GraphDataTools
 from utils import TEST_DATA_PATH_OBJECT, gen3_admin_tasks
+
+from gen3.auth import Gen3Auth
 
 logger = get_logger(__name__, log_level=os.getenv("LOG_LEVEL", "info"))
 
@@ -108,6 +110,6 @@ def create_program_project(user="main_account"):
     """
     Creates program and project if not present already
     """
-    api_key_json, auth_header = get_api_key_and_auth_header(user=user)
-    create_program("jnkns", auth_header, user)
-    create_project("jnkns", "jenkins", auth_header, user)
+    auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
+    sd_tools = GraphDataTools(auth=auth, program_name="jnkns", project_code="jenkins")
+    sd_tools.create_program_and_project()
